// src/app/core/services/feature-detection/feature-detection.enum.ts
export enum AppType {
    Unknown = 'unknown',
    Cordova = 'cordova',
    Electron = 'electron',
    Web = 'web'
  }
  
  export enum DeviceType {
    Unknown = 'unknown',
    Desktop = 'desktop',
    Tablet = 'tablet',
    Phone = 'phone'
  }
  
  export enum FeatureSystem {
    IPhone = 'iPhone',
    IOs = 'iOS',
    Safari = 'Safari',
    Chrome = 'Chrome',
    Android = 'Android',
    IPad = 'iPad',
    IPod = 'iPod',
    Windows = 'Windows',
    MacOS = 'MacOS',
    Unix = 'UNIX',
    Linux = 'Linux'
  }